+++ src/Constants.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#define USE_SOFTWARESERIAL
+
+/*
+
+
+
+*/