+++ src/Helpers/Logger.cpp
@@ -14,9 +14,9 @@ void Logger::setLogging(bool enable){
 }
 
 size_t Logger::write(uint8_t character) {
-  if (_loggingEnabled) _printer->write(character);
+  if (_loggingEnabled && _printer) _printer->write(character);
 }
 
 size_t Logger::write(const uint8_t* buffer, size_t size) {
-  if (_loggingEnabled) _printer->write(buffer, size);
+  if (_loggingEnabled && _printer) _printer->write(buffer, size);
 }