+++ src/Optolink.cpp
@@ -1,25 +1,8 @@
 #include "Optolink.h"
 
-/*
-const char error0 [] PROGMEM = "succes";
-const char error1 [] PROGMEM = "error1";
-const char error2 [] PROGMEM = "error2";
-const char error3 [] PROGMEM = "error3";
-const char error4 [] PROGMEM = "error4";
-const char error5 [] PROGMEM = "error5";
-
-const char* const errorMessages[] PROGMEM = {
-  error0,
-  error1,
-  error2,
-  error3,
-  error4,
-  error5
-};
-*/
 
 Optolink::Optolink():
-_serial(nullptr),
+_serialptr(nullptr),
 _address(0),
 _length(0),
 _value{0},
@@ -27,37 +10,40 @@ _writeMessageType(false),
 _rcvBuffer{0},
 _rcvBufferLen(0),
 _rcvLen(0),
-_timeoutTimer(0),
-_lastMillis(0),
-_errorCount(0),
-_sendMessage(false),
 _debugMessage(true),
 _state(RESET),
+_action(PROCESS),
+_lastMillis(0),
+_numberOfTries(5),
 _errorCode(0),
 _debugPrinter(nullptr)
 {}
 
 
 //begin serial @ 4800 baud, 8 bits, even parity, 2 stop bits
-void Optolink::begin(HardwareSerial* serial) {
-  _serial = serial;
-  _serial->begin(4800, SERIAL_8E2);
+#ifdef USE_SOFTWARESERIAL
+void Optolink::begin(const uint8_t rx, const uint8_t tx) {
+  _serialptr = new SoftwareSerial((int)rx, (int)tx, false, 64);
+  _serialptr->begin(4800);
+  _serialptr->setParity(PARITY_MODE_EVEN);
+  _serialptr->setStopBits(STOP_BITS_TWO);
+  //serial->flush();
+  //serial->swap();  //use GPIO15 and GPIO13 for TX/RX
 }
+#else
+void Optolink::begin(HardwareSerial* serial) {
+  _serialptr = serial;
+  _serialptr->begin(4800, SERIAL_8E2);
+  //serial->flush();
+  //serial->swap();  //use GPIO15 and GPIO13 for TX/RX
+}
+#endif
 
 
 void Optolink::loop() {
-  if (millis() - _timeoutTimer > 6 * 60 * 1000UL) {
-    //timeout in 6 minutes, reset
-    _state = RESET;
-    _debugPrinter->print(F("connection reset: timeout"));
-    return;
-  }
-  if (_errorCount >= 5 ) {
-    //5 errors, reset
-    _state = RESET;
-    _errorCount = 0;
-    _debugPrinter->print(F("connection reset: max errors exceeded."));
-    return;
+  if (_numberOfTries < 1) {
+    _state = IDLE;
+    _action = RETURN_ERROR;
   }
   switch (_state) {
     case RESET:
@@ -90,9 +76,6 @@ void Optolink::loop() {
     case RECEIVE:
       _receiveHandler();
       break;
-    case RETURN:
-      _returnHandler();
-      break;
   }
 }
 
@@ -100,21 +83,21 @@ void Optolink::loop() {
 //Set communication with Vitotronic to defined state = reset to KW protocol
 void Optolink::_resetHandler() {
   const uint8_t buff[] = {0x04};
-  _serial->write(buff, sizeof(buff));
+  _serialptr->write(buff, sizeof(buff));
   _lastMillis = millis();
   _state = RESET_ACK;
   if (_debugMessage) {
     _debugPrinter->println(F("Resetting Optolink..."));
     _debugMessage = false;
   }
-  return;
 }
+
+
 void Optolink::_resetAckHandler() {
-  if (_serial->available()) {
-    if (_serial->peek() == 0x05) { //use peek so connection can be made immediately in next state
+  if (_serialptr->available()) {
+    if (_serialptr->peek() == 0x05) {  //use peek so connection can be made immediately in next state
       //received 0x05/enquiry: optolink has been reset
       _state = INIT;
-      _timeoutTimer = millis();
       _debugMessage = true;
       _debugPrinter->println(F("Optolink reset."));
     }
@@ -123,7 +106,7 @@ void Optolink::_resetAckHandler() {
     }
   }
   else {
-    if (millis() - _lastMillis > 500) {
+    if (millis() - _lastMillis > 500) {  //try again every 0,5sec
       _state = RESET;
     }
   }
@@ -136,26 +119,25 @@ void Optolink::_initHandler() {
     _debugPrinter->println(F("Establishing Optolink connection..."));
     _debugMessage = false;
   }
-  if (_serial->available()) {
-    if (_serial->read() == 0x05) {
+  if (_serialptr->available()) {
+    if (_serialptr->read() == 0x05) {
       //0x05/ENQ received, sending initiator
       const uint8_t buff[] = {0x16, 0x00, 0x00};
-      _serial->write(buff, sizeof(buff));
+      _serialptr->write(buff, sizeof(buff));
       _lastMillis = millis();
-      _timeoutTimer = _lastMillis;
       _state = INIT_ACK;
     }
   }
-  return;
 }
+
+
 void Optolink::_initAckHandler() {
-  if (_serial->available()) {
-    if (_serial->read() == 0x06) {
+  if (_serialptr->available()) {
+    if (_serialptr->read() == 0x06) {
       //ACK received, moving to next state
       _state = IDLE;
-      _timeoutTimer = millis();
+      _action = WAIT;
       _debugMessage = true;
-      _errorCount = 0;
       //debug: done
       _debugPrinter->println(F("Optolink connection established."));
     }
@@ -165,51 +147,42 @@ void Optolink::_initAckHandler() {
       _state = INIT;
     }
   }
-  return;
+  if (millis() - _lastMillis > 10 * 1000UL) {  //if no ACK is coming, reset connection
+    _state = RESET;
+  }
 }
 
 
 //idle state, waiting for user action
 void Optolink::_idleHandler() {
-  if (millis() - _lastMillis > 2 * 60 * 1000UL) {
-    //send SYNC every 2 minutes to keep communication alive
+  if (millis() - _lastMillis > 2 * 60 * 1000UL) {  //send SYNC every 2 minutes to keep communication alive
     _state = SYNC;
   }
   _clearInputBuffer(); //keep input clean
-  if (_sendMessage) _state = SYNC;
-  else _state = IDLE;
-  return;
+  if (_action == PROCESS) _state = SYNC;
 }
 
 
 //send SYNC (= initiator)
 void Optolink::_syncHandler() {
-  if (_debugMessage) {
-    _debugPrinter->println(F("Syncing Optolink..."));
-    _debugMessage = false;
-  }
   const uint8_t buff[] = {0x16, 0x00, 0x00};
-  _serial->write(buff, sizeof(buff));
-  _state = SYNC_ACK;
+  _serialptr->write(buff, sizeof(buff));
   _lastMillis = millis();
-  return;
+  _state = SYNC_ACK;
 }
 
 
 void Optolink::_syncAckHandler() {
-  if (_serial->available()) {
-    if (_serial->read() == 0x06) {
-      _timeoutTimer = millis();
-      _debugMessage = true;
-      _debugPrinter->println(F("Optolink synced."));
-      if(_sendMessage) _state = SEND;
+  if (_serialptr->available()) {
+    if (_serialptr->read() == 0x06) {
+      if(_action == PROCESS) _state = SEND;
       else _state = IDLE;
     }
-    else {
-      _errorCount++;
-    }
   }
-  return;
+  if (millis() - _lastMillis > 2 * 1000UL) {  //if no ACK is coming, reset connection
+    _state = RESET;
+    if (_action = PROCESS) --_numberOfTries;
+  }
 }
 
 
@@ -228,7 +201,7 @@ void Optolink::_sendHandler() {
     //add value to message
     memcpy(&buff[7], _value, 1);
     buff[7 + _length] = _calcChecksum(buff, 7 + _length);
-    _serial->write(buff, 8 + _length);
+    _serialptr->write(buff, 8 + _length);
   }
   else {
     //type is READ
@@ -243,132 +216,108 @@ void Optolink::_sendHandler() {
     buff[7] = _calcChecksum(buff, 8);
     //set length of expected answer
     _rcvLen = 8 + _length;
-    _serial->write(buff, 8);
+    _serialptr->write(buff, 8);
   }
-  _lastMillis = millis();
-  _sendMessage = false;
   _rcvBufferLen = 0;
+  _lastMillis = millis();
+  --_numberOfTries;
   _state = SEND_ACK;
-  _debugPrinter->print(F("Message sent for "));
   if (_writeMessageType) _debugPrinter->print(F("WRITE "));
-  else _debugPrinter->print(F("READ "));
-  _debugPrinter->print(F("on address "));
+  else _debugPrinter->print(F("READ"));
+  _debugPrinter->print(F(" request on address "));
   _debugPrinter->print(_address, HEX);
-  _debugPrinter->print(F(" with length "));
-  _debugPrinter->print(_length);
-  _debugPrinter->println(F("."));
-  return;
+  _debugPrinter->print(F(", length "));
+  _debugPrinter->println(_length);
 }
 
 
 void Optolink::_sendAckHandler() {
-  if (_serial->available()) {
-    uint8_t buff = _serial->read();
-    _timeoutTimer = millis();
-    if (buff == 0x06) {
-      //transmit succesful, moving to next state
+  if (_serialptr->available()) {
+    uint8_t buff = _serialptr->read();
+    if (buff == 0x06) {  //transmit succesful, moving to next state
+      _debugPrinter->println(F("req: ack"));
       _state = RECEIVE;
-      _debugPrinter->println(F("Message send ACK."));
     }
-    else if (buff == 0x15) {
-      //transmit negatively acknowledged, return to SYNC and try again
-      _timeoutTimer = millis();
-      _sendMessage = true;
-      _errorCount++;
+    else if (buff == 0x15) {  //transmit negatively acknowledged, return to SYNC and try again
+      _debugPrinter->println(F("req: nack"));
       _state = SYNC;
       _clearInputBuffer();
-      _debugPrinter->println(F("Message send NACK, trying again"));
     }
   }
-  else if (millis() - _timeoutTimer > 2 * 1000UL) {
-    //timeout waiting for ACK 2secs, return to SYNC and try again
-    _sendMessage = true;
+  if (millis() - _lastMillis > 2 * 1000UL) {  //if no ACK is coming, return to SYNC and try again
+    _debugPrinter->println(F("req: t/o"));
     _state = SYNC;
-    _errorCount++;
     _clearInputBuffer();
-    _debugPrinter->println(F("Message send timeout, trying again"));
   }
-  return;
 }
 
 
 void Optolink::_receiveHandler() {
-  while (_serial->available() > 0) {  //while instead of if: read complete RX buffer
-    _rcvBuffer[_rcvBufferLen] = _serial->read();
+  while (_serialptr->available() > 0) {  //while instead of if: read complete RX buffer
+    _rcvBuffer[_rcvBufferLen] = _serialptr->read();
     if (_rcvBuffer[0] != 0x41) return; //find out why this is needed! I'd expect the rx-buffer to be empty.
-    _rcvBufferLen++;
-    _timeoutTimer = millis();
+    ++_rcvBufferLen;
   }
-  if (_rcvBufferLen == _rcvLen) {
-    //message complete, check/decode
-    _debugPrinter->println(F("Message received, checking."));
-    if (_rcvBuffer[1] != (_rcvLen - 3)) {
-      _errorCount++;
-      _sendMessage = false;
-      _state = RETURN;
-      _errorCode = 2;
-      _debugPrinter->println(F("Length check error, skipping command"));
+  if (_rcvBufferLen == _rcvLen) {  //message complete, check message
+    if (_rcvBuffer[1] != (_rcvLen - 3)) {  //check for message length
+      _numberOfTries = 0;
+      _errorCode = 1;
+      _debugPrinter->println("ERRORCODE4");
       return;
     }
-    if (!_checkChecksum(_rcvBuffer, _rcvLen)) {
-      //checksum is wrong, trying again
-      _errorCount++;
-      _sendMessage = true;
+    if (_rcvBuffer[2] != 0x01) {  //Vitotronic returns an error message, skipping DP
+      _numberOfTries = 0;
+      _errorCode = 2;  //Vitotronic error
+      return;
+    }
+    if (!_checkChecksum(_rcvBuffer, _rcvLen)) {  //checksum is wrong, trying again
       _rcvBufferLen = 0;
+      _errorCode = 3;  //checksum error
       memset(_rcvBuffer, 0, 12);
       _state = SYNC;
-      _debugPrinter->println(F("Checksum error, trying again"));
-      return;
-    }
-    if (_rcvBuffer[2] != 0x01) {
-      //Vitotronic returns a failure message, skipping DP
-      _errorCount++;
-      _sendMessage = false;
-      _state = RETURN;
-      _errorCode = 3;
-      _debugPrinter->println(F("command NOK, skipping command"));
       return;
     }
     if (_rcvBuffer[3] == 0x01) {
-      //message is response from READ, so value is returned correctly
-      //ESP8266 is little endian so least significant byte first
-      //debug: Succes
-      //old solution: _value.byte4Value = (_rcvBuffer[10] << 24) | (_rcvBuffer[9] << 16) | (_rcvBuffer[8] << 8) | _rcvBuffer[7];
-      }
-      _state = RETURN;
+      //message is from READ command, so returning read value
     }
-  return;
-}
-
-
-void Optolink::_returnHandler() {
-  //stay in this state untill message has been read
-  //waiting for read()-action
+    _state = IDLE;
+    _action = RETURN;
+    _errorCode = 0;  //succes
+    _debugPrinter->println(F("succes"));
+    return;
+  }
+  if (millis() - _lastMillis > 10 * 1000UL) {  //Vitotronic isn't answering, try again
+    _rcvBufferLen = 0;
+    _errorCode = 1;  //Connection error
+    memset(_rcvBuffer, 0, 12);
+    _state = SYNC;
+  }
 }
 
 
 //set properties for datapoint and move state to SEND
 bool Optolink::readFromDP(uint16_t address, uint8_t length) {
-  if (_state != IDLE && (_state == SYNC && _sendMessage)) {
-    _debugPrinter->println(F("Previous process not finished, skipping this action"));
+  if (_action != WAIT) {
+    _debugPrinter->println(F("Optolink available, skipping action."));
     return false;
   }
   //setup properties for next state in communicationHandler
   _address = address;
   _length = length;
   _writeMessageType = false;
-  _sendMessage = true;
   _rcvBufferLen = 0;
+  _numberOfTries = 5;
   memset(_rcvBuffer, 0, 12);
-  _state = SYNC;
+  _state = SYNC;  //avoid an extra loop to pass by idleHandler
+  _action = PROCESS;
   return true;
 }
 
 
 //set properties datapoint and move state to SEND
 bool Optolink::writeToDP(uint16_t address, uint8_t length, uint8_t value[]) {
-  if (_state != IDLE && (_state == SYNC && _sendMessage)) {
-    _debugPrinter->println(F("Previous process not finished, skipping this action"));
+  if (_action != WAIT) {
+    _debugPrinter->println(F("Optolink available, skipping action."));
     return false;
   }
   //setup variables for next state
@@ -376,50 +325,65 @@ bool Optolink::writeToDP(uint16_t address, uint8_t length, uint8_t value[]) {
   _length = length;
   memcpy(_value, value, _length);
   _writeMessageType = true;
-  _sendMessage = true;
   _rcvBufferLen = 0;
+  _numberOfTries = 5;
   memset(_rcvBuffer, 0, 12);
-  _state = SYNC;
+  _state = SYNC;  //avoid an extra loop to pass by idleHandler
+  _action = PROCESS;
   return true;
 }
 
 
-const int8_t Optolink::available() {
-  if (_errorCode) return -1;
-  else if (_state == RETURN) return 1;
+const int8_t Optolink::available() const {
+  if (_action == RETURN_ERROR) return -1;
+  else if (_action == RETURN) return 1;
   else return 0;
+}
+
+const bool Optolink::isBusy() const {
+  if (_action == WAIT) return false;
+  else return true;
 }
 
 
 //return value and reset comunication to IDLE
 void Optolink::read(uint8_t value[]) {
-  if (_state != RETURN) {
-    _debugPrinter->println(F("Optolink warning, read value before it is available."));
+  //ESP8266 is little endian so least significant byte first
+  //old solution: value = (_rcvBuffer[10] << 24) | (_rcvBuffer[9] << 16) | (_rcvBuffer[8] << 8) | _rcvBuffer[7];
+  if (_action != RETURN) {
+    _debugPrinter->println(F("No reading available"));
+    return;
   }
   if (_writeMessageType) {  //return original value in case of WRITE command
     memcpy(value, &_value, _length);
-    _state = IDLE;
+    _action = WAIT;
     return;
   }
   else {
     memcpy(value, &_rcvBuffer[7], _length);
-    _state = IDLE;
-    return;
+    _action = WAIT;
+    return;  //added for clarity
   }
+}
+
+
+const uint8_t Optolink::readError() {
+  _action = WAIT;
+  return _errorCode;
 }
 
 
 //calculate Checksum
 inline uint8_t Optolink::_calcChecksum(uint8_t array[], uint8_t length) {
   uint8_t sum = 0;
-  for (uint8_t i = 1; i < length - 1 ; i++){  //start with second byte and en before checksum
+  for (uint8_t i = 1; i < length - 1 ; ++i){  //start with second byte and en before checksum
     sum += array[i];
   }
   return sum;
 }
 inline bool Optolink::_checkChecksum(uint8_t array[], uint8_t length) {
   uint8_t sum = 0;
-  for (uint8_t i = 1; i < length - 1; i++) { //start with second byte and en before checksum
+  for (uint8_t i = 1; i < length - 1; ++i) { //start with second byte and en before checksum
     sum += array[i];
   }
   return (array[length - 1] == sum);
@@ -428,8 +392,8 @@ inline bool Optolink::_checkChecksum(uint8_t array[], uint8_t length) {
 
 //clear serial input buffer
 inline void Optolink::_clearInputBuffer() {
-  while(_serial->available() > 0) {
-    _serial->read();
+  while(_serialptr->available() > 0) {
+    _serialptr->read();
   }
 }
 
@@ -437,27 +401,22 @@ void Optolink::setDebugPrinter(Print* printer) {
   _debugPrinter = printer;
 }
 
-void Optolink::readError(char* errorString, uint8_t length) const  {
-  //copy errorMessage to errorString, keeping an eye on length
-  strncpy(errorString, "this is an errorMessage", length - 1);
-  return;
-};
 
 //Copied from Arduino.cc forum --> (C) robtillaart
 inline void Optolink::_printHex83(uint8_t array[], uint8_t length) {
   char tmp[length * 2 + 1];
   byte first;
   uint8_t j = 0;
-  for (uint8_t i = 0; i < length; i++) {
+  for (uint8_t i = 0; i < length; ++i) {
     first = (array[i] >> 4) | 48;
     if (first > 57) tmp[j] = first + (byte)39;
     else tmp[j] = first ;
-    j++;
+    ++j;
 
     first = (array[i] & 0x0F) | 48;
     if (first > 57) tmp[j] = first + (byte)39;
     else tmp[j] = first;
-    j++;
+    ++j;
   }
   tmp[length * 2] = 0;
   _debugPrinter->print(tmp);