+++ src/Optolink.h
@@ -1,22 +1,33 @@
 #pragma once
 #include <Arduino.h>
-#include "Helpers/Datatypes.h"
+#include "Constants.h"
+#ifdef USE_SOFTWARESERIAL
+#include <SoftwareSerial.h>
+#endif
 
 class Optolink {
-
   public:
     Optolink();
+#ifdef USE_SOFTWARESERIAL
+    void begin(const uint8_t rx, const uint8_t tx);
+#else
     void begin(HardwareSerial* serial);
+#endif
     void loop();
-    const int8_t available();
+    const int8_t available() const;
+    const bool isBusy() const;
     bool readFromDP(uint16_t address, uint8_t length);
     bool writeToDP(uint16_t address, uint8_t length, uint8_t value[]);
     void read(uint8_t value[]);
-    void readError(char* errorString, uint8_t length) const;
+    const uint8_t readError();
     void setDebugPrinter(Print* printer);
 
   private:
-    HardwareSerial* _serial;
+#ifdef USE_SOFTWARESERIAL
+    SoftwareSerial* _serialptr;
+#else
+    HardwareSerial* _serialptr;
+#endif
     enum OptolinkState: uint8_t {
       RESET,
       RESET_ACK,
@@ -27,9 +38,14 @@ class Optolink {
       SYNC_ACK,
       SEND,
       SEND_ACK,
-      RECEIVE,
-      RETURN
+      RECEIVE
     } _state;
+    enum OptolinkAction: uint8_t {
+      WAIT,
+      PROCESS,
+      RETURN,
+      RETURN_ERROR
+    } _action;
     uint16_t _address;
     uint8_t _length;
     bool _writeMessageType;
@@ -37,11 +53,9 @@ class Optolink {
     uint8_t _rcvBuffer[12];
     uint8_t _rcvBufferLen;
     uint8_t _rcvLen;
-    uint32_t _timeoutTimer;
     uint32_t _lastMillis;
-    uint8_t _errorCount;
+    uint8_t _numberOfTries;
     uint8_t _errorCode;
-    bool _sendMessage;
     void _resetHandler();
     void _resetAckHandler();
     void _initHandler();